Use the useRef in parent component.
